Split SlackInfoCommander out of SlackCommander
